Move AreYouSureModal to separate modal module
