+++ src/app/elitefour/modals/index.ts
@@ -1,0 +1,1 @@
+export * from './are-you-sure-modal/are-you-sure-modal.component';
