+++ src/app/elitefour/modals/modal.module.ts
@@ -1,0 +1,11 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {AreYouSureModalComponent} from './are-you-sure-modal/are-you-sure-modal.component';
+
+@NgModule({
+  declarations: [AreYouSureModalComponent],
+  imports: [CommonModule],
+  exports: []
+})
+export class ModalModule {
+}
